+++ CHANGELOG.md
@@ -49,6 +49,7 @@ a flexibility update.
 - bugfix: keep "empty" top bar visible in the menu.
 - bugfix: set NSRequiresAquaSystemAppearance to false in /Applications/Notion.app/Contents/Info.plist
   so system dark/light mode can be properly detected.
+- bugfix: make ctrl+f popover shadow less extreme.
 - tweak: sticky table/list rows.
 - theme: "material ocean" = an oceanic colour palette.
 - theme: "dracula" = a theme based on the popular dracula color palette